moved SystemInfo class into system_info.py
